Refactor PdfViewer into modular components and fix critical issues
- **Modular Architecture**: Split 722-line PdfViewer.tsx into focused components
  - LoadingState: Loading display with progress bar
  - ErrorState: Error message display
  - EmptyState: Initial PDF selection screen
  - PdfCanvas: Core PDF rendering and interaction logic (400+ lines)
  - Main PdfViewer: Simple orchestrator component (50 lines)

- **Critical Bug Fixes**:
  - Fix PDF.js canvas concurrent rendering error with proper task cancellation
  - Fix canvas size calculation using window dimensions instead of getBoundingClientRect
  - Fix zoom operation after reset by properly clearing stale scale values
  - Fix passive event listener error for mouse wheel zoom with non-passive listener

- **Technical Improvements**:
  - Add rendering race condition prevention with isRenderingRef
  - Implement proper cleanup for PDF render tasks
  - Optimize zoom state management in useZoom hook
  - Maintain all ARIA accessibility attributes
  - Preserve responsive design and touch interactions

- **Testing**: 13 focused test cases covering core functionality
- **Performance**: Improved maintainability while preserving all features
