+++ src/components/PdfViewer/index.ts
@@ -1,0 +1,1 @@
+export { PdfViewer } from './PdfViewer';
